+++ src/mousedb/animal/urls/strain.py
@@ -35,9 +35,13 @@ urlpatterns = patterns('',
 		'login_required':True,
 		'template_name':'confirm_delete.html'
 		}, name="strain-delete"),
-    url(r'^(?P<strain_slug>[\w-]+)/(?P<background_slug>[\w-]+)/all/?$', views.StrainBackgroundDetailAll.as_view(), name="strain-background-detail"),
-    url(r'^(?P<strain_slug>[\w-]+)/(?P<background_slug>[\w-]+)/?$', views.StrainBackgroundDetail.as_view(), name="strain-background-detail-all"),    
+    url(r'background/new/?$', views.BackgroundCreate.as_view(), name="background-new"),
+    url(r'background/(?P<background_slug>[\w-]+)/all/?$', views.BackgroundDetail.as_view(), name="background-detail-all"),
+    url(r'background/(?P<slug>[\w-]+)/edit/?$', views.BackgroundUpdate.as_view(), name="background-edit"),   
+    url(r'background/(?P<slug>[\w-]+)/delete/?$', views.BackgroundDelete.as_view(), name="background-delete"),      
+    url(r'background/(?P<background_slug>[\w-]+)/?$', views.BackgroundDetail.as_view(), name="background-detail"),
+    url(r'^(?P<slug>[\w-]+)/(?P<background_slug>[\w-]+)/all/?$', views.StrainBackgroundDetailAll.as_view(), name="strain-background-detail-all"),
+    url(r'^(?P<slug>[\w-]+)/(?P<background_slug>[\w-]+)/?$', views.StrainBackgroundDetail.as_view(), name="strain-background-detail"),    
     url(r'^(?P<slug>[\w-]+)/all/?$', views.StrainDetailAll.as_view(), name="strain-detail-all"),
     url(r'^(?P<slug>[\w-]+)/?$', views.StrainDetail.as_view(), name="strain-detail"),
-    url(r'background/(?P<background_slug>[\w-]+)/?$', views.BackgroundDetail.as_view(), name="background-detail")
 )